Refactor: agregar nuevos componentes y actualizar rutas para el cuidado prenatal, laboral y postnatal
