+++ src/maternal-and-child-health/components/labour-delivery/partography/labour-delivery.scss
@@ -1,6 +1,6 @@
 @use '@carbon/styles/scss/spacing';
 @use '@carbon/styles/scss/type';
-@import '../../root.scss';
+@import '../../../../root.scss';
 
 .widgetContainer {
   background-color: $ui-background;
