+++ src/maternal-and-child-health/components/postnatal-care/postpartumControl.component.tsx
@@ -15,7 +15,7 @@ import {
 } from '@carbon/react';
 import { launchPatientWorkspace, CardHeader, EmptyState } from '@openmrs/esm-patient-common-lib';
 import { useConfig, useLayoutType } from '@openmrs/esm-framework';
-import styles from './prenatalCareChart.scss';
+import styles from './postnatalCareChart.scss';
 import dayjs from 'dayjs';
 import { usePostpartumControlTable } from '../../../hooks/usePostpartumControl';
 import type { ConfigObject } from '../../../config-schema';
